requirements.txt
requests
beautifulsoup4
pymongo
python-dotenv
streamlit
snscrape
